Move files from 'data' to 'src'
