+++ scripts/src/voucherify-setup/uninstall.ts
